DEV: Modularize files in lib
