--- lib/discourse_translator/post_extension.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module DiscourseTranslator
-  module PostExtension
-    extend ActiveSupport::Concern
-    prepended { before_update :clear_translations, if: :raw_changed? }
-    include Translatable
-  end
-end
