--- lib/discourse_translator/topic_extension.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module DiscourseTranslator
-  module TopicExtension
-    extend ActiveSupport::Concern
-    prepended { before_update :clear_translations, if: :title_changed? }
-    include Translatable
-  end
-end
